fix(tests): separate constructor tests from integration tests
- Create HangfireSchemaInitializerUnitTests for constructor validation
- Remove Docker-dependent setup from simple constructor tests
- Tests now pass without requiring TestContainers infrastructure
- Maintains test coverage while improving test reliability
